+++ doc/Changelog.md
@@ -1,3 +1,9 @@
+### version v4.9.15 (Jan 3, 2021)
+
+* fixed downloading: only one copy of every media
+* fixed addition of subalbum to selection
+* bug fixes
+
 ### version v4.9.14 (Jan 2, 2021)
 
 * fix bugs